requirements.txt
Django==2.2.4
pytest-django==3.5.1
django-rest-swagger==2.2.0
djangorestframework==3.9.4
pytest-cov==2.7.1
coverage==4.5.4
croniter==0.3.30
tornado==6.0.3
whitenoise==3.0
graphene-django==2.9.0
django-filter==2.2.0
django-graphql-jwt==0.1.5